[TASK] Create PortalUser automatically, adapt templates
PortalUser instances are created automatically for users if
none exists to make sure every account has a party.

All links to content now have a blank target and twitter footers
are linked to the tweet status.

The first two content tiles are now always sponsored content regardless
of the age of the post.